Adjustment to styles and some file restructuring
