+++ components/Accordion/Accordion.module.scss
@@ -27,6 +27,7 @@
 
     .seasonEnded {
       margin-left: auto;
+      margin-right: 3px;
       text-align: right;
       font-size: 1rem;
       font-weight: normal;
@@ -41,7 +42,7 @@
 
 .accordionContent {
   padding-top: 0.25rem;
-  margin-left: 2rem;
+  margin: 0 1rem;
   overflow: auto hidden;
   transition: height 400ms cubic-bezier(0.77, 0, 0.175, 1);
 }
