+++ pages/winners/Winners.module.scss
@@ -1,0 +1,6 @@
+.winnerAccordions {
+  display: flex;
+  flex-direction: column;
+  gap: 12px;
+  margin-top: 0.5rem;
+}
