+++ pages/winners/index.ts
@@ -1,0 +1,1 @@
+export { default, getStaticProps } from "./Winners";
