+++ styles/globals.scss
@@ -43,7 +43,7 @@ hr {
 main {
   margin: auto;
   padding: 1rem 3rem;
-  max-width: clamp(30rem, 100%, 50rem);
+  max-width: clamp(30rem, 100%, 60rem);
   @include responsive {
     padding: 1rem;
     max-width: none;
